requirements.txt
Flask>=3.1.2
flask-cors>=6.0.2
pillow>=10.2.0,<11.0.0
pdf2image>=1.16.0
torch>=2.0.0
transformers>=4.40.0,<4.50.0
accelerate>=0.25.0
bitsandbytes>=0.43.0
surya-ocr==0.8.2
python-dotenv>=1.0.0
httpx>=0.24.0
supabase==2.4.0
pypdfium2==4.30.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
